requirements.txt
python-telegram-bot==21.0.1
python-dotenv==1.0.1
fastf1==3.8.1